third_party/WebKit/LayoutTests/editing/pasteboard/copy-paste-ruby-text-expected.txt

This tests copying and pasting text with ruby doesn't insert new lines before and after rt.To manually test, copy and paste the text into the textarea. There should be not new line in the textarea.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textarea.value.indexOf('\n') is -1
PASS successfullyParsed is true

TEST COMPLETE
